+++ temperature/print.h
@@ -29,7 +29,7 @@
 
 void printString(const char *s);
 void printPgmString(const char *s);
-void printInteger(long n);
+void printInteger(uint32_t n);
 void print_uint8_base2(uint8_t n);
 void printFloat(float n);
